[FEAT]: ✨ baseTableViewVC 분리
